+++ source/isaaclab/isaaclab/terrains/utils.py
@@ -11,6 +11,9 @@ import torch
 import trimesh
 import warp as wp
 
+from pxr import UsdGeom
+
+import isaaclab.sim as sim_utils
 from isaaclab.utils.warp import raycast_mesh
 
 
@@ -78,11 +81,6 @@ def create_prim_from_mesh(prim_path: str, mesh: trimesh.Trimesh, **kwargs):
         visual_material: The visual material to apply. Defaults to None.
         physics_material: The physics material to apply. Defaults to None.
     """
-    # need to import these here to prevent isaacsim launching when importing this module
-    from pxr import UsdGeom
-
-    import isaaclab.sim as sim_utils
-
     # create parent prim
     sim_utils.create_prim(prim_path, "Xform")
     # create mesh prim
