+++ source/isaaclab_experimental/isaaclab_experimental/envs/__init__.py
@@ -42,14 +42,6 @@ For more information about the workflow design patterns, see the `Task Design Wo
 .. _`Task Design Workflows`: https://isaac-sim.github.io/IsaacLab/source/features/task_workflows.html
 """
 
-from .direct_rl_env_warp import DirectRLEnvWarp  # noqa: F401
-from .interactive_scene_warp import InteractiveSceneWarp  # noqa: F401
-from .manager_based_env_warp import ManagerBasedEnvWarp  # noqa: F401
-from .manager_based_rl_env_warp import ManagerBasedRLEnvWarp  # noqa: F401
+from isaaclab.utils.module import lazy_export
 
-__all__ = [
-    "DirectRLEnvWarp",
-    "InteractiveSceneWarp",
-    "ManagerBasedEnvWarp",
-    "ManagerBasedRLEnvWarp",
-]
+lazy_export()
