+++ source/isaaclab_experimental/isaaclab_experimental/envs/mdp/__init__.pyi
@@ -1,0 +1,29 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+# Forward all stable MDP terms (commands/curriculums/events/observations/...) via a
+# lazy fallback, so unresolved names defer to the stable package without eagerly
+# importing its backend-dependent submodules.
+from isaaclab.envs.mdp import *  # noqa: F401, F403
+
+# Experimental Warp-first action terms. Listed by name (not ``*``) so the runtime
+# implementations in ``joint_actions`` stay lazy and only the pure-data config
+# classes are imported when an env config is constructed.
+from .actions import (  # noqa: F401
+    JointAction,
+    JointActionCfg,
+    JointEffortAction,
+    JointEffortActionCfg,
+    JointPositionAction,
+    JointPositionActionCfg,
+)
+
+# Override stable terms with experimental Warp-first implementations. These leaf
+# modules are import-clean (no eager backend imports), so re-exporting them here
+# is safe.
+from .events import *  # noqa: F401, F403
+from .observations import *  # noqa: F401, F403
+from .rewards import *  # noqa: F401, F403
+from .terminations import *  # noqa: F401, F403
