+++ source/isaaclab_experimental/isaaclab_experimental/envs/mdp/actions/__init__.pyi
@@ -1,0 +1,16 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "JointAction",
+    "JointActionCfg",
+    "JointEffortAction",
+    "JointEffortActionCfg",
+    "JointPositionAction",
+    "JointPositionActionCfg",
+]
+
+from .actions_cfg import JointActionCfg, JointEffortActionCfg, JointPositionActionCfg
+from .joint_actions import JointAction, JointEffortAction, JointPositionAction
