+++ source/isaaclab_experimental/isaaclab_experimental/managers/__init__.pyi
@@ -1,0 +1,51 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "ActionManager",
+    "ActionTerm",
+    "CommandManager",
+    "CommandTerm",
+    "EventManager",
+    "ManagerBase",
+    "ManagerTermBase",
+    "ActionTermCfg",
+    "CommandTermCfg",
+    "CurriculumTermCfg",
+    "EventTermCfg",
+    "ManagerTermBaseCfg",
+    "ObservationGroupCfg",
+    "ObservationTermCfg",
+    "RecorderTermCfg",
+    "RewardTermCfg",
+    "TerminationTermCfg",
+    "ObservationManager",
+    "RewardManager",
+    "SceneEntityCfg",
+    "TerminationManager",
+]
+
+from isaaclab.managers import *  # noqa: F401, F403
+
+from .action_manager import ActionManager, ActionTerm
+from .command_manager import CommandManager, CommandTerm
+from .event_manager import EventManager
+from .manager_base import ManagerBase, ManagerTermBase
+from .manager_term_cfg import (
+    ActionTermCfg,
+    CommandTermCfg,
+    CurriculumTermCfg,
+    EventTermCfg,
+    ManagerTermBaseCfg,
+    ObservationGroupCfg,
+    ObservationTermCfg,
+    RecorderTermCfg,
+    RewardTermCfg,
+    TerminationTermCfg,
+)
+from .observation_manager import ObservationManager
+from .reward_manager import RewardManager
+from .scene_entity_cfg import SceneEntityCfg
+from .termination_manager import TerminationManager
