+++ source/isaaclab_experimental/isaaclab_experimental/managers/scene_entity_cfg.py
@@ -11,11 +11,15 @@ with the stable manager stack (which type-checks against the stable SceneEntityC
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import warp as wp
 
-from isaaclab.assets.articulation.base_articulation import BaseArticulation
+from isaaclab.assets import ArticulationCfg
 from isaaclab.managers.scene_entity_cfg import SceneEntityCfg as _SceneEntityCfg
-from isaaclab.scene import InteractiveScene
+
+if TYPE_CHECKING:
+    from isaaclab.scene import InteractiveScene
 
 
 class SceneEntityCfg(_SceneEntityCfg):
@@ -41,7 +45,7 @@ class SceneEntityCfg(_SceneEntityCfg):
         entity = scene[self.name]
 
         # -- Warp joint mask / ids for articulations
-        if isinstance(entity, BaseArticulation):
+        if isinstance(entity.cfg, ArticulationCfg):
             if self.joint_ids == slice(None):
                 joint_ids_list = list(range(entity.num_joints))
                 mask_list = [True] * entity.num_joints
