+++ source/isaaclab_tasks/changelog.d/fix-warp-env-cfg-forbidden-imports.skip
@@ -1,0 +1,2 @@
+Test-only: extend test_env_cfg_no_forbidden_imports to also scan
+isaaclab_tasks_experimental tasks. No user-facing change.
