+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/direct/ant/__init__.py
@@ -20,7 +20,7 @@ gym.register(
     entry_point=f"{__name__}.ant_env_warp:AntWarpEnv",
     disable_env_checker=True,
     kwargs={
-        "env_cfg_entry_point": f"{__name__}.ant_env_warp:AntWarpEnvCfg",
+        "env_cfg_entry_point": f"{__name__}.ant_env_warp_cfg:AntWarpEnvCfg",
         "rl_games_cfg_entry_point": f"{stable_agents}:rl_games_ppo_cfg.yaml",
         "rsl_rl_cfg_entry_point": f"{stable_agents}.rsl_rl_ppo_cfg:AntPPORunnerCfg",
         "skrl_cfg_entry_point": f"{stable_agents}:skrl_ppo_cfg.yaml",
