+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/direct/ant/ant_env_warp.py
@@ -5,83 +5,9 @@
 
 from __future__ import annotations
 
-from isaaclab_newton.physics import MJWarpSolverCfg, NewtonCfg
-
-import isaaclab.sim as sim_utils
-from isaaclab.assets import ArticulationCfg
-from isaaclab.envs import DirectRLEnvCfg
-from isaaclab.scene import InteractiveSceneCfg
-from isaaclab.sim import SimulationCfg
-from isaaclab.terrains import TerrainImporterCfg
-from isaaclab.utils.configclass import configclass
-
-from isaaclab_assets import ANT_CFG
-
 from isaaclab_tasks_experimental.direct.locomotion.locomotion_env_warp import LocomotionWarpEnv
 
-
-@configclass
-class AntWarpEnvCfg(DirectRLEnvCfg):
-    # env
-    episode_length_s = 15.0
-    decimation = 2
-    action_scale = 0.5
-    action_space = 8
-    observation_space = 36
-    state_space = 0
-
-    solver_cfg = MJWarpSolverCfg(
-        njmax=45,
-        nconmax=25,
-        cone="pyramidal",
-        integrator="implicitfast",
-        impratio=1,
-    )
-    newton_cfg = NewtonCfg(
-        solver_cfg=solver_cfg,
-        num_substeps=1,
-        debug_mode=False,
-        use_cuda_graph=True,
-    )
-
-    # simulation
-    sim: SimulationCfg = SimulationCfg(dt=1 / 120, render_interval=decimation, physics=newton_cfg)
-    terrain = TerrainImporterCfg(
-        prim_path="/World/ground",
-        terrain_type="plane",
-        collision_group=-1,
-        physics_material=sim_utils.RigidBodyMaterialCfg(
-            friction_combine_mode="average",
-            restitution_combine_mode="average",
-            static_friction=1.0,
-            dynamic_friction=1.0,
-            restitution=0.0,
-        ),
-        debug_vis=False,
-    )
-
-    # scene
-    scene: InteractiveSceneCfg = InteractiveSceneCfg(
-        num_envs=4096, env_spacing=4.0, replicate_physics=True, clone_in_fabric=True
-    )
-
-    # robot
-    robot: ArticulationCfg = ANT_CFG.replace(prim_path="/World/envs/env_.*/Robot")
-    joint_gears: list = [15, 15, 15, 15, 15, 15, 15, 15]
-
-    heading_weight: float = 0.5
-    up_weight: float = 0.1
-
-    energy_cost_scale: float = 0.05
-    actions_cost_scale: float = 0.005
-    alive_reward_scale: float = 0.5
-    dof_vel_scale: float = 0.2
-
-    death_cost: float = -2.0
-    termination_height: float = 0.31
-
-    angular_velocity_scale: float = 1.0
-    contact_force_scale: float = 0.1
+from .ant_env_warp_cfg import AntWarpEnvCfg
 
 
 class AntWarpEnv(LocomotionWarpEnv):
