+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/classic/cartpole/mdp/__init__.pyi
@@ -1,0 +1,14 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "joint_pos_target_l2",
+]
+
+# Forward stable MDP terms and experimental Warp-first overrides lazily, then
+# override with cartpole-specific terms below.
+from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+
+from .rewards import joint_pos_target_l2
