+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/classic/humanoid/mdp/__init__.py
@@ -5,7 +5,6 @@
 
 """This sub-module contains the functions that are specific to the humanoid environments."""
 
-from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+from isaaclab.utils.module import lazy_export
 
-from .observations import *  # noqa: F401, F403
-from .rewards import *  # noqa: F401, F403
+lazy_export()
