+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/classic/humanoid/mdp/__init__.pyi
@@ -1,0 +1,29 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "base_yaw_roll",
+    "base_up_proj",
+    "base_heading_proj",
+    "base_angle_to_target",
+    "upright_posture_bonus",
+    "move_to_target_bonus",
+    "progress_reward",
+    "joint_pos_limits_penalty_ratio",
+    "power_consumption",
+]
+
+# Forward stable MDP terms and experimental Warp-first overrides lazily, then
+# override with humanoid-specific terms below.
+from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+
+from .observations import base_angle_to_target, base_heading_proj, base_up_proj, base_yaw_roll
+from .rewards import (
+    joint_pos_limits_penalty_ratio,
+    move_to_target_bonus,
+    power_consumption,
+    progress_reward,
+    upright_posture_bonus,
+)
