+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/classic/humanoid/mdp/rewards.py
@@ -20,11 +20,11 @@ from isaaclab_experimental.managers.manager_base import ManagerTermBase
 from isaaclab_newton.kernels.state_kernels import rotate_vec_to_body_frame
 
 import isaaclab.utils.string as string_utils
-from isaaclab.assets import Articulation
 
 if TYPE_CHECKING:
     from isaaclab_experimental.managers.manager_term_cfg import RewardTermCfg
 
+    from isaaclab.assets import Articulation
     from isaaclab.envs import ManagerBasedRLEnv
