+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/locomotion/velocity/mdp/__init__.pyi
@@ -1,0 +1,30 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "terrain_levels_vel",
+    "feet_air_time",
+    "feet_air_time_positive_biped",
+    "feet_slide",
+    "stand_still_joint_deviation_l1",
+    "track_ang_vel_z_world_exp",
+    "track_lin_vel_xy_yaw_frame_exp",
+    "terrain_out_of_bounds",
+]
+
+# Forward stable MDP terms and experimental Warp-first overrides (commands, actions,
+# observations, ...) lazily, then override with locomotion-specific terms below.
+from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+
+from .curriculums import terrain_levels_vel
+from .rewards import (
+    feet_air_time,
+    feet_air_time_positive_biped,
+    feet_slide,
+    stand_still_joint_deviation_l1,
+    track_ang_vel_z_world_exp,
+    track_lin_vel_xy_yaw_frame_exp,
+)
+from .terminations import terrain_out_of_bounds
