+++ source/isaaclab_tasks_experimental/isaaclab_tasks_experimental/manager_based/manipulation/reach/mdp/__init__.pyi
@@ -1,0 +1,16 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+__all__ = [
+    "position_command_error",
+    "position_command_error_tanh",
+    "orientation_command_error",
+]
+
+# Forward stable MDP terms and experimental Warp-first overrides lazily, then
+# override with reach-specific terms below.
+from isaaclab_experimental.envs.mdp import *  # noqa: F401, F403
+
+from .rewards import orientation_command_error, position_command_error, position_command_error_tanh
